requirements.txt
asgiref==3.5.2
attrs==21.4.0
autobahn==22.5.1
Automat==20.2.0
autopep8==1.6.0
certifi==2022.6.15
cffi==1.15.0
channels==3.0.5
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
constantly==15.1.0
cryptography==37.0.2
daphne==3.0.2
dj-database-url==0.5.0
Django==3.2.13
django-environ==0.8.1
djangorestframework==3.13.1
filelock==3.7.1
gunicorn==20.1.0
huggingface-hub==0.8.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
joblib==1.1.0
numpy==1.23.0
packaging==21.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pyOpenSSL==22.0.0
pyparsing==3.0.9
pytz==2021.3
PyYAML==6.0
regex==2022.6.2
requests==2.28.0
sacremoses==0.0.53
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tokenizers==0.12.1
toml==0.10.2
torch==1.11.0
tqdm==4.64.0
transformers==4.18.0
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.2.0
urllib3==1.26.9
whitenoise==5.3.0
zope.interface==5.4.0
